RRVector::size() returns size_t.
+compatibility with std::vector, later we can switch
-printf symbol for size_t is %zd, not usual %d, this creates new class of errors
 (wxString::Format crashes without 'z',
  _vsnprintf in RRReporter::report seems to be ok with both, but i added 'z' anyway)

+++ include/Lightsprint/RRVector.h
@@ -161,7 +161,7 @@ public:
 		}
 	}
 	//! Returns number of elements in vector (like std::vector).
-	unsigned size() const
+	size_t size() const
 	{
 		return numUsed;
 	}
